fix: use env with oidc configs
