+++ package.json
@@ -74,6 +74,8 @@
     "husky": "^8.0.3",
     "lint-staged": "^15.2.0",
     "prettier": "^3.1.1",
+    "remark": "^15.0.1",
+    "remark-cli": "^12.0.0",
     "stylelint": "^15",
     "typescript": "^5"
   },
