+++ src/config/oidc.mjs
@@ -2,11 +2,11 @@
 export default {
   AUTH_DATA: 'authData',
   server: {
-    url: 'https://portal.172.22.96.136.nip.io',
+    url: process.env.OIDC_SERVER_URL,
   },
   client: {
-    client_id: 'bff-client',
-    client_secret: '61324af0-1234-4f61-b110-ef57013267d6',
-    redirect_uri: '/oidc/callback',
+    client_id: process.env.CLIENT_ID,
+    client_secret: process.env.CLIENT_SECRET,
+    redirect_uri: process.env.CLIENT_REDIRECT_URI,
   },
 };
